// src/ts/enums/UniversityImage.ts
export enum UniversityImage {
    upc = "upc",
    euncet = "euncet",
    uic = "uic",
    ub = "ub",
    cei = "cei",
    cett = "cett",
    enti = "enti",
    escac = "escac",
    esrp = "esrp",
    escodi = "escodi",
    ispc = "ispc",
    uniba = "uniba",
    eina = "eina",
    escolamassana = "escolamassana",
    eui = "eui",
    gimbernat = "gimbernat",
    euss = "euss",
    euit = "euit",
    guttmann = "guttmann",
    uab = "uab",
    upf = "upf",
    esci = "esci",
    tecnocampus = "tecnocampus",
    udl = "udl",
    eurl = "eurl",
    inefc = "inefc",
    ostelea = "ostelea",
    next = "next",
    uvic = "uvic",
    udg = "udg",
    euses = "euses",
    euroaula = "euroaula",
    euht = "euht",
    eram = "eram",
    eufb = "eufb",
    mediterrani = "mediterrani",
    blanquerna = "blanquerna",
    iqs = "iqs",
    lasalle = "lasalle",
    filo = "filo",
    esade = "esade",
    pere = "pere",
    bioetica = "bioetica",
    esdi = "esdi",
    eatm = "eatm",
    abat_oliba = "abat_oliba",
    eshob = "eshob",
    idep = "idep",
    santotomas = "santotomas",
    urv = "urv",
    cesda = "cesda",
    igema = "igema",
    iaac = "iaac",
    harbour = "harbour",
    escrbcc = "escrbcc",
    esic = "esic",
    iese = "iese",
    gbsb = "gbsb",
    llotja = "llotja",
    liceu = "liceu",
    eae = "eae",
    uibs = "uibs",
    davinci = "davinci",
}